requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
torch==2.6.0
transformers==4.49.0
